+++ CHANGELOG.md
@@ -9,6 +9,8 @@ Includes major fixes and improvements contributed by [@lollox80](https://github.
 - Overrides now respect `message` or `title` in the `data` section for channel specific messages
   - New [recipe](docs/recipes/channel_specific_messages.md) to illustrate it
 - Overrides now match against the name of a transport if there's no matching delivery
+- Delivery overrides that have indirect targets like `person` now correctly resolved to email / sms / slack / mobile etc
+- Explicitly enabled deliveries in action call not overridden by scenarios
 
 ## 1.11.1
 Includes fixes and improvements contributed by [@lollox80](https://github.com/lollox80)